refactor: move code related to description formatter to separate file
